# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DoD AI Tool Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/{toolId}/launch:
    post:
      tags: [Launch]
      operationId: launchTool
      summary: Initiate a tool launch (issues a single-use launch token)
      parameters:
        - name: toolId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Launch initiated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LaunchInitiateResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /tools/context-exchange:
    post:
      tags: [Launch]
      operationId: exchangeContextToken
      summary: Tool exchanges a launch token for the user's context bundle and a session token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContextExchangeRequest' }
      responses:
        '200':
          description: Context bundle
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContextExchangeResponse' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /tools/library-query:
    post:
      tags: [Launch]
      operationId: queryLibrary
      summary: Tool queries the user's library with a session token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LibraryQueryRequest' }
      responses:
        '200':
          description: Snippets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryQueryResponse' }
        '401':
          description: Invalid session token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /launches/recent:
    get:
      tags: [Launch]
      operationId: listRecentLaunches
      summary: List the user's recent tool launches
      responses:
        '200':
          description: Launches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LaunchHistoryItem' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ---------- Admin ----------
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    UserProfile:
      type: object
      required:
        [userId, branch, rank, mosCode, dutyTitle, unit, baseLocation,
         securityClearance, deploymentStatus, primaryMission,
         aiUseCases, freeFormContext, isAdmin, completenessPct,
         contextBlock, activePresetId, updatedAt]
      properties:
        userId: { type: string }
        branch: { type: [string, 'null'] }
        rank: { type: [string, 'null'] }
        mosCode: { type: [string, 'null'] }
        dutyTitle: { type: [string, 'null'] }
        unit: { type: [string, 'null'] }
        baseLocation: { type: [string, 'null'] }
        securityClearance: { type: [string, 'null'] }
        deploymentStatus: { type: [string, 'null'] }
        primaryMission: { type: [string, 'null'] }
        aiUseCases:
          type: array
          items: { type: string }
        freeFormContext: { type: [string, 'null'] }
        isAdmin: { type: boolean }
        completenessPct: { type: integer }
        contextBlock: { $ref: '#/components/schemas/ContextBlockState' }
        activePresetId: { type: [string, 'null'] }
        updatedAt: { type: string, format: date-time }
...
    ContextBlockScores:
      type: object
      required: [doctrine, environment, constraints, experience]
      properties:
        doctrine: { type: integer, description: Criterion 1 score (1-3). }
        environment:
          type: integer
          description: Criterion 2 — environment & commander's intent (1-3).
        constraints:
          type: integer
          description: Criterion 3 — constraints, limitations & risk (1-3).
        experience: { type: integer, description: Criterion 4 score (1-3). }
    ContextBlockEvaluation:
      type: object
      required: [submissionId, scores, totalScore, status, opsecFlag, flags]
      properties:
        submissionId: { type: string }
        scores: { $ref: '#/components/schemas/ContextBlockScores' }
        totalScore:
          type: integer
          description: Sum of the four criteria, /12. Forced to 0 on OPSEC violations.
        status:
          type: string
          description: '"GO" when totalScore >= 10 and OPSEC clean, else "NO-GO".'
        opsecFlag:
          type: boolean
          description: True if the submission tripped the OPSEC fail-safe.
        flags:
          type: string
          description: Brief description of OPSEC violation or ambiguity, "None" otherwise.
    ContextBlockState:
      type: object
      required:
        [doctrine, intent, environment, constraints, risk, experience,
         confirmedAt, lastEvaluation]
      properties:
        doctrine: { type: [string, 'null'] }
        intent: { type: [string, 'null'] }
        environment: { type: [string, 'null'] }
        constraints: { type: [string, 'null'] }
        risk: { type: [string, 'null'] }
        experience: { type: [string, 'null'] }
        confirmedAt:
          type: [string, 'null']
          format: date-time
          description: When the operator last accepted this Context Block.
        lastEvaluation:
          oneOf:
            - $ref: '#/components/schemas/ContextBlockEvaluation'
            - type: 'null'
...
    LaunchInitiateResponse:
      type: object
      required: [launchId, launchToken, launchUrl, expiresAt]
      properties:
        launchId: { type: string }
        launchToken: { type: string }
        launchUrl: { type: string }
        expiresAt: { type: string, format: date-time }
    ContextExchangeRequest:
      type: object
      required: [launchToken]
      properties:
        launchToken: { type: string, minLength: 1 }
    ContextExchangeResponse:
      type: object
      required:
        [sessionToken, sessionExpiresAt, tool, user, profile, contextBlock,
         structuredContextBlock, primer]
      properties:
        sessionToken: { type: string }
        sessionExpiresAt: { type: string, format: date-time }
        tool:
          type: object
          required: [id, slug, name, vendor, atoStatus]
          properties:
            id: { type: string }
            slug: { type: string }
            name: { type: string }
            vendor: { type: string }
            atoStatus: { type: string }
        user:
          type: object
          required: [id, displayName, email]
          properties:
            id: { type: string }
            displayName: { type: string }
            email: { type: [string, 'null'] }
        profile: { $ref: '#/components/schemas/UserProfile' }
        contextBlock:
          type: string
          description: >
            Pre-formatted Markdown block summarizing the user's identity,
            mission, assignment, and (when confirmed) the 6-element
            Context Block. Tool builders can drop this directly into
            their model prompt without re-parsing the structured profile.
        structuredContextBlock:
          oneOf:
            - $ref: '#/components/schemas/ContextBlockState'
            - type: 'null'
          description: >
            Structured 6-element Context Block as last confirmed by the
            operator, or null when they have not yet confirmed one.
        primer: { $ref: '#/components/schemas/RagPrimer' }
    RagPrimer:
      type: object
      required: [queries, snippets]
      properties:
        queries:
          type: array
          items: { type: string }
        snippets:
          type: array
          items: { $ref: '#/components/schemas/RagSnippet' }
    RagSnippet:
      type: object
      required:
        [chunkId, documentId, documentTitle, chunkIndex, content, score]
      properties:
        chunkId: { type: string }
        documentId: { type: string }
        documentTitle: { type: string }
        chunkIndex: { type: integer }
        content: { type: string }
        score: { type: number }
    LibraryQueryRequest:
      type: object
      required: [sessionToken, query]
      properties:
        sessionToken: { type: string, minLength: 1 }
        query: { type: string, minLength: 1 }
        limit: { type: integer }
    LibraryQueryResponse:
      type: object
      required: [query, snippets]
      properties:
        query: { type: string }
        snippets:
          type: array
          items: { $ref: '#/components/schemas/RagSnippet' }
...
    LaunchHistoryItem:
      type: object
      required: [id, toolId, toolName, toolSlug, status, createdAt]
      properties:
        id: { type: string }
        toolId: { type: string }
        toolName: { type: string }
        toolSlug: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
...